Fix timeouts on failed tests when running in Mocha
When using waitForUpdate under Mocha, test failure results in timeout as Mocha `done` does not use `done.fail(e)`. Instead, Mocha just calls `done(e)` with the error. Update to try that if `done.fail` is not present.
Additionally fixes the slight documentation issue with an incomplete chain complaining that `then(done)` is missing when it should be `end(done)`.